refactor(core,auth): move OAuth2 store traits to core::auth
Moves `ClientStore`, `AuthCode`+`AuthCodeStore`, `StoredRefreshToken`+
`RefreshTokenStore`, and `DeviceState`+`DeviceCodeStore` into
`assistant_core::auth`. The auth crate keeps the in-memory
implementations and re-exports the moved names for back-compat. The
SQLite implementations in `assistant_storage::auth_state_store` now
import these contracts from core, removing one chunk of the
storage→auth coupling.

Refs openspec/changes/resolve-priority-tech-debt/specs/storage-layering/spec.md
